Fix Barclayscrape for new (~Q1 2015) design
This is actually much nicer overall, all credit to Barclays' frontend
engineers. Nonetheless, it was tricky enough that barclayscrape has been
broken for several months, and I've been manually downloading statements
daily.

The main issues were a surprising and difficult-to-debug switch to using
user-agent-based browser sniffing, plus the significant changes in the
frontend causing some issues.

Concerningly, there is still a fair amount of perplexing server-side state
which means that you can end up with an OFX file which identifies itself as
being for account A but has the transactions from account B.

To fix this, we *must* make the appropriate menu clicks in the UI, even though
we still go ahead and do the download using casper.download.